requirements.txt
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
DateTime==5.5
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
logger==1.4
MarkupSafe==3.0.1
numpy==2.1.2
paramiko==3.5.0
pillow==10.4.0
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==75.1.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
urllib3==2.2.3
waitress==3.0.0
Werkzeug==3.0.4
zope.interface==7.1.0
